Move internal things to internal
